+++ services/supervisor/src/lib.rs
@@ -81,7 +81,7 @@ use exonum::{
     crypto::Hash,
     runtime::{
         rust::{
-            api::ServiceApiBuilder, AfterCommitContext, Broadcaster, BuiltinInstance, CallContext,
+            api::ServiceApiBuilder, AfterCommitContext, Broadcaster, CallContext, DefaultInstance,
             Service,
         },
         BlockchainData, InstanceId, SUPERVISOR_INSTANCE_ID,
@@ -358,7 +358,7 @@ where
     }
 }
 
-impl<Mode: mode::SupervisorMode> BuiltinInstance for Supervisor<Mode> {
+impl<Mode: mode::SupervisorMode> DefaultInstance for Supervisor<Mode> {
     const INSTANCE_ID: u32 = SUPERVISOR_INSTANCE_ID;
     const INSTANCE_NAME: &'static str = Self::NAME;
 }