+++ services/supervisor/tests/simple.rs
@@ -21,7 +21,7 @@ use exonum::{
     helpers::{Height, ValidatorId},
     messages::{AnyTx, Verified},
     runtime::{
-        rust::{BuiltinInstance, CallContext, Service},
+        rust::{CallContext, DefaultInstance, Service},
         ArtifactId, BlockchainData, DispatcherError, ExecutionError, InstanceId, SnapshotExt,
     },
 };
@@ -56,7 +56,7 @@ pub fn sign_config_propose_transaction_by_us(
 #[service_factory(artifact_name = "config-change-test-service")]
 pub struct ConfigChangeService;
 
-impl BuiltinInstance for ConfigChangeService {
+impl DefaultInstance for ConfigChangeService {
     const INSTANCE_ID: InstanceId = 119;
     const INSTANCE_NAME: &'static str = "config-change";
 }
@@ -145,7 +145,7 @@ fn change_consensus_config_with_one_confirmation() {
 
     let mut testkit = TestKitBuilder::auditor()
         .with_validators(initial_validator_count)
-        .with_builtin_rust_service(SimpleSupervisor::new())
+        .with_default_rust_service(SimpleSupervisor::new())
         .create();
 
     let cfg_change_height = Height(5);
@@ -192,8 +192,8 @@ fn change_consensus_config_with_one_confirmation() {
 fn service_config_change() {
     let mut testkit = TestKitBuilder::validator()
         .with_validators(2)
-        .with_builtin_rust_service(SimpleSupervisor::new())
-        .with_builtin_rust_service(ConfigChangeService)
+        .with_default_rust_service(SimpleSupervisor::new())
+        .with_default_rust_service(ConfigChangeService)
         .create();
 
     let cfg_change_height = Height(5);
@@ -223,8 +223,8 @@ fn service_config_change() {
 fn incorrect_actual_from_field() {
     let mut testkit = TestKitBuilder::validator()
         .with_validators(2)
-        .with_builtin_rust_service(SimpleSupervisor::new())
-        .with_builtin_rust_service(ConfigChangeService)
+        .with_default_rust_service(SimpleSupervisor::new())
+        .with_default_rust_service(ConfigChangeService)
         .create();
 
     let cfg_change_height = Height(5);
@@ -250,7 +250,7 @@ fn incorrect_actual_from_field() {
 fn discard_config_propose_from_auditor() {
     let mut testkit = TestKitBuilder::auditor()
         .with_validators(2)
-        .with_builtin_rust_service(SimpleSupervisor::new())
+        .with_default_rust_service(SimpleSupervisor::new())
         .create();
 
     let cfg_change_height = Height(5);
@@ -297,7 +297,7 @@ fn discard_config_propose_from_auditor() {
 fn test_send_proposal_with_api() {
     let mut testkit = TestKitBuilder::validator()
         .with_validators(2)
-        .with_builtin_rust_service(SimpleSupervisor::new())
+        .with_default_rust_service(SimpleSupervisor::new())
         .create();
 
     let old_validators = testkit.network().validators();
@@ -351,7 +351,7 @@ fn test_send_proposal_with_api() {
 #[test]
 fn deploy_service() {
     let mut testkit = TestKitBuilder::validator()
-        .with_builtin_rust_service(SimpleSupervisor::new())
+        .with_default_rust_service(SimpleSupervisor::new())
         .with_rust_service(DeployableService)
         .create();
 
@@ -393,7 +393,7 @@ fn actual_from_is_zero() {
 
     let mut testkit = TestKitBuilder::auditor()
         .with_validators(initial_validator_count)
-        .with_builtin_rust_service(SimpleSupervisor::new())
+        .with_default_rust_service(SimpleSupervisor::new())
         .create();
 
     // Change height set to 0