+++ services/supervisor/tests/supervisor/main.rs
@@ -216,7 +216,7 @@ fn start_service_instance(testkit: &mut TestKit, instance_name: &str) -> Instanc
 fn testkit_with_inc_service() -> TestKit {
     TestKitBuilder::validator()
         .with_logger()
-        .with_builtin_rust_service(DecentralizedSupervisor::new())
+        .with_default_rust_service(DecentralizedSupervisor::new())
         .with_rust_service(IncService)
         .create()
 }
@@ -224,7 +224,7 @@ fn testkit_with_inc_service() -> TestKit {
 fn testkit_with_inc_service_and_n_validators(n: u16) -> TestKit {
     TestKitBuilder::validator()
         .with_logger()
-        .with_builtin_rust_service(DecentralizedSupervisor::new())
+        .with_default_rust_service(DecentralizedSupervisor::new())
         .with_rust_service(IncService)
         .with_validators(n)
         .create()
@@ -237,7 +237,7 @@ fn testkit_with_inc_service_and_two_validators() -> TestKit {
 fn testkit_with_inc_service_auditor_validator() -> TestKit {
     TestKitBuilder::auditor()
         .with_logger()
-        .with_builtin_rust_service(DecentralizedSupervisor::new())
+        .with_default_rust_service(DecentralizedSupervisor::new())
         .with_rust_service(IncService)
         .with_validators(1)
         .create()
@@ -246,8 +246,8 @@ fn testkit_with_inc_service_auditor_validator() -> TestKit {
 fn testkit_with_inc_service_and_static_instance() -> TestKit {
     TestKitBuilder::validator()
         .with_logger()
-        .with_builtin_rust_service(DecentralizedSupervisor::new())
-        .with_builtin_rust_service(IncService)
+        .with_default_rust_service(DecentralizedSupervisor::new())
+        .with_default_rust_service(IncService)
         .create()
 }
 
@@ -553,7 +553,7 @@ fn test_start_two_services_in_one_request() {
 fn test_restart_node_and_start_service_instance() {
     let mut testkit = TestKitBuilder::validator()
         .with_logger()
-        .with_builtin_rust_service(DecentralizedSupervisor::new())
+        .with_default_rust_service(DecentralizedSupervisor::new())
         .with_rust_service(IncService)
         .create();
     deploy_default(&mut testkit);
@@ -1024,7 +1024,7 @@ fn test_id_assignment_sparse() {
     // Create testkit with builtin instance with ID 100.
     let mut testkit = TestKitBuilder::validator()
         .with_logger()
-        .with_builtin_rust_service(DecentralizedSupervisor::new())
+        .with_default_rust_service(DecentralizedSupervisor::new())
         .with_artifact(inc_service_artifact.clone())
         .with_instance(inc_service_artifact.into_instance(max_builtin_id, "inc"))
         .with_rust_service(inc_service)